Int grid transform generation

+++ src/assets.rs
@@ -19,7 +19,7 @@ fn ldtk_path_to_asset_path<'a, 'b>(
         .into()
 }
 
-pub type TilesetMap = HashMap<i64, Handle<Texture>>;
+pub type TilesetMap = HashMap<i32, Handle<Texture>>;
 
 #[derive(TypeUuid)]
 #[uuid = "ecfb87b7-9cd9-4970-8482-f2f68b770d31"]
